refactor: initialise project from template
Follow the template TODOs to create the actions service.

TIS21-5582
TIS21-5585
