+++ .aws/task-definition-template.json
@@ -1,31 +1,31 @@
 {
   "containerDefinitions": [
     {
-      "name": "tis-template",
-      "image": "430723991443.dkr.ecr.eu-west-2.amazonaws.com/tis-template:latest",
+      "name": "tis-trainee-actions",
+      "image": "430723991443.dkr.ecr.eu-west-2.amazonaws.com/tis-trainee-actions:latest",
       "secrets": [
         {
           "name": "SENTRY_DSN",
-          "valueFrom": "tis-template-sentry-dsn"
+          "valueFrom": "tis-trainee-actions-sentry-dsn"
         }
       ],
       "logConfiguration": {
         "logDriver": "awslogs",
         "options": {
-          "awslogs-group": "awslogs-${environment}-tis-template",
+          "awslogs-group": "awslogs-${environment}-tis-trainee-actions",
           "awslogs-region": "eu-west-2",
-          "awslogs-stream-prefix": "awslogs-tis-template"
+          "awslogs-stream-prefix": "awslogs-tis-trainee-actions"
         }
       },
       "portMappings": [
         {
-          "containerPort": 8080
+          "containerPort": 8212
         }
       ],
       "environment": [
         {
           "name": "TITLE",
-          "value": "tis-template"
+          "value": "tis-trainee-actions"
         },
         {
           "name": "AWS_REGION",
@@ -38,7 +38,7 @@
       ]
     }
   ],
-  "family": "tis-template-${environment}",
+  "family": "tis-trainee-actions-${environment}",
   "requiresCompatibilities": [
     "FARGATE"
   ],
