+++ build.gradle.kts
@@ -9,8 +9,7 @@ plugins {
   id("org.sonarqube") version "4.4.1.3373"
 }
 
-// TODO: Update group to end with "admin" or "trainee".
-group = "uk.nhs.tis"
+group = "uk.nhs.tis.trainee"
 version = "0.0.1"
 
 configurations {
@@ -61,8 +60,7 @@ sonarqube {
     property("sonar.host.url", "https://sonarcloud.io")
     property("sonar.login", System.getenv("SONAR_TOKEN"))
     property("sonar.organization", "health-education-england")
-    // TODO: Update sonar.projectKey to real value.
-    property("sonar.projectKey", "Health-Education-England_tis-microservice-template")
+    property("sonar.projectKey", "Health-Education-England_tis-trainee-actions")
 
     property("sonar.java.checkstyle.reportPaths",
       "build/reports/checkstyle/main.xml,build/reports/checkstyle/test.xml")
