+++ settings.gradle.kts
@@ -1,2 +1,1 @@
-// TODO: Update project name.
-rootProject.name = "tis-microservice-template"
+rootProject.name = "tis-trainee-actions"
