+++ src/main/java/uk/nhs/tis/trainee/actions/TisTraineeActionsApplication.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright 2022 Crown Copyright (Health Education England)
+ * Copyright 2024 Crown Copyright (Health Education England)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
  * associated documentation files (the "Software"), to deal in the Software without restriction,
@@ -19,16 +19,15 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package uk.nhs.hee.tis.template;
+package uk.nhs.tis.trainee.actions;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-//TODO: Update package and class name.
-public class TemplateApplication {
+public class TisTraineeActionsApplication {
 
   public static void main(String[] args) {
-    SpringApplication.run(TemplateApplication.class);
+    SpringApplication.run(TisTraineeActionsApplication.class);
   }
 }
