+++ src/main/resources/application.yml
@@ -1,9 +1,11 @@
+application:
+  environment: ${ENVIRONMENT:local}
+
 server:
-  # TODO: Update server port and context path to a unique value.
-  port: 8080
+  port: 8212
   servlet:
-    context-path: /template
+    context-path: /actions
 
 sentry:
   dsn: ${SENTRY_DSN:}
-  environment: ${SENTRY_ENVIRONMENT:}
+  environment: ${application.environment}
